Add files via upload
This is not as percentage as I wished, completed yet. Mostly it is recoding some things to achieve somewhat correct syntax. Now I need to reduce the useless statements not applicable to the project1. Those are the sections that are not used in the restaurant program. 

I am treating it a paper, constant refinement into the preferred final.