Add files via upload
Just revising as I go through the code.  It also helps me reconstruct in my mind what is occurring between the various POM, DOM, Objects, Servlets and others. 
The syntax of the code along with the writing it to process is a challenge. The set and get are definitely needed to keep it more stable.