Sunday redrafting of code to review and adapt and apply. More was removed to lewave only relevant to the project1. I continue to revise the classes and methods. this. access the instance of the class of a specific object.
The menu items must be change by the Admin. Just like the order can change.
That requires similar code of editing the datasource. Create a method which has a selection to enable modifying the data tables of items on the menu stored in the database(managed in SQL via Azure).